outsourced dex
